requirements.txt
flask
pymysql
flask-sqlalchemy
flask-script
flask-wtf
python-dotenv
flask-migrate